+++ engine/conftest.py
@@ -429,7 +429,7 @@ def get_user_permission_role_mapping_from_frontend_plugin_json() -> RoleMapping:
     for role in plugin_json["roles"]:
         if grants := role["grants"]:
             for permission in role["role"]["permissions"]:
-                # only concerned with grafana-oncall-app specific grants
+                # only concerned with pando-oncall-app specific grants
                 # ignore things like plugins.app:access actions
                 action = permission["action"]
                 permission_class = None