+++ grafana-plugin/e2e-tests/utils/clients/grafana.ts
@@ -130,7 +130,7 @@ class GrafanaAPIClient {
   };
 
   makeRequest = async (request: APIRequestContext, path: string, method: 'get' | 'post' | 'put' | 'delete' = 'post'): Promise<APIResponse> => {
-    const res = await request[method](`${BASE_URL}/api/plugins/grafana-oncall-app/${path.replace(/^\//, '')}`, {
+    const res = await request[method](`${BASE_URL}/api/plugins/pando-oncall-app/${path.replace(/^\//, '')}`, {
       headers: this.requestHeaders,
     });
     expect(res.ok()).toBeTruthy();